Added flat icons and support for them
These are used in the list of store categories on the new store view controller.  This involved adding an “image_flat” attribute to the Core Data schema and CloudKit schema and uploading images for them to CloudKit.  This also rippled through other parts of them model class and other classes.