Fixed the duplicates for updates problem
I was creating new instances of types and then calling update in the NotificationController.  I added update methods on the type ModelControllers that take records and update the associated Core Data record.

I also started making changes in preparation for the UI refreshes, which I’ll finish in the next commit.